Alterando o tipo da variavel ID que estava em conflito com System32

+++ PSI/Models/Categoria.cs
@@ -7,7 +7,7 @@ namespace PSI.Models
 {
     public class Categoria
     {
-        public long CategoriaId { get; set; }
+        public int CategoriaId { get; set; }
         public string Nome { get; set; }
     }
 }
